fix(example): move command into own folders
