+++ src/Command/GenerateFibonacci/GenerateXFibonacciCommand.php
@@ -1,21 +1,20 @@
 <?php
 
-namespace App\Command;
+namespace App\Command\GenerateFibonacci;
 
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\Stopwatch\Stopwatch;
 
 #[AsCommand(
-    name: 'generate:fibonacci',
-    description: 'Generate a Fibonacci suite with maximum iteration from array.',
+    name: 'generate:x-fibonacci',
+    description: 'Generate a Fibonacci suite with maximum iteration from Generator.',
 )]
-class GenerateFibonacciCommand extends Command
+final class GenerateXFibonacciCommand extends Command
 {
     protected function configure(): void
     {
@@ -31,7 +30,7 @@ class GenerateFibonacciCommand extends Command
         $stopwatch = new Stopwatch();
 
         $io->title('Generation started...');
-        $stopwatch->start('fibonacci');
+        $stopwatch->start('x-fibonacci');
 
         if ($input->getArgument('max')) {
             $fibonacciSuite = $this->fibonacci((int) $input->getArgument('max'));
@@ -42,7 +41,7 @@ class GenerateFibonacciCommand extends Command
         foreach ($fibonacciSuite as $number) {
             $io->write(sprintf('%d ', $number));
         }
-        $event = $stopwatch->stop('fibonacci');
+        $event = $stopwatch->stop('x-fibonacci');
 
         $io->success(
             sprintf('The fibonacci suite is generated with success. %s', $event->__toString())
@@ -51,20 +50,17 @@ class GenerateFibonacciCommand extends Command
         return Command::SUCCESS;
     }
 
-    private function fibonacci(int $maxIter = 10): iterable
+    private function fibonacci(?int $maxIter = 10): iterable
     {
         $current = 1;
         $previous = 0;
-        $return = [];
 
         for ($iter = 0; $iter < $maxIter; $iter++) {
             $temp = $current;
             $current = $previous + $current;
             $previous = $temp;
 
-            $return[] = $current;
+            yield $current;
         }
-
-        return $return;
     }
 }
